make detail page more visually appealing and some fixes
